Update webapp to add docker healthcheck
The HEALTHCHECK instruction tells Docker how to test a container to check that it is still working.
This can detect cases such as a web server that is stuck in an infinite loop and unable to handle new connections, even though the server process is still running.

source: https://docs.docker.com/engine/reference/builder/